==> 1_Data_Labelling.ipynb <==
from google.colab import drive
drive.mount('/content/gdrive/')

import numpy as np
import pandas as pd

traffic_df = pd.read_csv('/content/gdrive/MyDrive/CIC-Darknet/Darknet.csv')
traffic_df.head()

traffic_df["Label.1"].loc[traffic_df["Label.1"] == "AUDIO-STREAMING"] = "Audio-Streaming"
traffic_df["Label.1"].loc[traffic_df["Label.1"] == "File-transfer"] = "File-Transfer"
traffic_df["Label.1"].loc[traffic_df["Label.1"] == "Video-streaming"] = "Video-Streaming"

traffic_df.groupby('Label')['Label.1'].value_counts()

### Non-Tor, Non-Vpn

non_tor = traffic_df['Label'] == 'Non-Tor'
non_vpn = traffic_df['Label'] == 'NonVPN'

label_benign = non_tor | non_vpn
benign_df = traffic_df[label_benign]
benign_df['Label.1'].value_counts()

benign_df["Label.1"].loc[benign_df["Label.1"] == "Audio-Streaming"] = "benign_Audio-Streaming"
benign_df["Label.1"].loc[benign_df["Label.1"] == "Chat"] = "benign_Chat"
benign_df["Label.1"].loc[benign_df["Label.1"] == "File-Transfer"] = "benign_File-Transfer"
benign_df["Label.1"].loc[benign_df["Label.1"] == "VOIP"] = "benign_VOIP"
benign_df["Label.1"].loc[benign_df["Label.1"] == "Video-Streaming"] = "benign_Video-Streaming"
benign_df["Label.1"].loc[benign_df["Label.1"] == "Email"] = "benign_Email"
benign_df["Label.1"].loc[benign_df["Label.1"] == "Browsing"] = "benign_Browsing"
benign_df["Label.1"].loc[benign_df["Label.1"] == "P2P"] = "benign_P2P"

benign_df['Label.1'].value_counts()

### Tor, Vpn

tor = traffic_df['Label'] == 'Tor'
vpn = traffic_df['Label'] == 'VPN'

label_darknet = tor | vpn
darknet_df = traffic_df[label_darknet]
darknet_df['Label.1'].value_counts()

darknet_df["Label.1"].loc[darknet_df["Label.1"] == "Audio-Streaming"] = "dark_Audio-Streaming"
darknet_df["Label.1"].loc[darknet_df["Label.1"] == "Chat"] = "dark_Chat"
darknet_df["Label.1"].loc[darknet_df["Label.1"] == "File-Transfer"] = "dark_File-Transfer"
darknet_df["Label.1"].loc[darknet_df["Label.1"] == "VOIP"] = "dark_VOIP"
darknet_df["Label.1"].loc[darknet_df["Label.1"] == "Video-Streaming"] = "dark_Video-Streaming"
darknet_df["Label.1"].loc[darknet_df["Label.1"] == "Email"] = "dark_Email"
darknet_df["Label.1"].loc[darknet_df["Label.1"] == "Browsing"] = "dark_Browsing"
darknet_df["Label.1"].loc[darknet_df["Label.1"] == "P2P"] = "dark_P2P"

darknet_df['Label.1'].value_counts()

data = pd.concat([benign_df, darknet_df])
data['Label.1'].value_counts()

data.reset_index(drop=True, inplace=True)
data.head()

data.to_csv('/content/gdrive/MyDrive/CIC-Darknet/darknet.csv', index=False)
