==> 2.Data Preprocessing.ipynb <==
from google.colab import drive
drive.mount('/content/gdrive/')

import numpy as np
import pandas as pd

traffic_df = pd.read_csv('/content/gdrive/MyDrive/CIC-Darknet/darknet.csv')
traffic_df.shape

### Drop Columns

drop_columns = ['Flow ID', 'Src IP', 'Dst IP', 'Timestamp', 'Label']
traffic_df.rename(columns={'Label.1':'target'}, inplace=True)
traffic_df.drop(drop_columns, axis=1, inplace=True)

traffic_df.head()

### Reduce Memory Usage

def reduce_mem_usage(df, verbose=True):
    numerics = ['int16', 'int32', 'int64', 'float16', 'float32', 'float64']
    start_mem = df.memory_usage().sum() / 1024**2
    for col in df.columns:
        col_type = df[col].dtypes
        if col_type in numerics:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)

    end_mem = df.memory_usage().sum() / 1024**2
    print('Memory usage after optimization is: {:.2f} MB'.format(end_mem))
    print('Decreased by {:.1f}%'.format(100 * (start_mem - end_mem) / start_mem))

    return df

traffic_df = reduce_mem_usage(traffic_df)

### Infinity and Null Value Replace 0

traffic_df = traffic_df.replace([np.inf, -np.inf], np.nan).dropna(axis=0)

columns = traffic_df.columns

for column in columns:
    print(pd.isnull(traffic_df[column]).sum())

### Label Encoding

traffic_df['target'].value_counts()

from sklearn.preprocessing import LabelEncoder

encoder = LabelEncoder()

traffic_df['target'] = encoder.fit_transform(traffic_df['target'])
traffic_df['target'].value_counts()

traffic_df.head()

### Normalization (Log Transform)

columns = []

columns += list(traffic_df.select_dtypes(['float16']).columns)
columns += list(traffic_df.select_dtypes(['float32']).columns)
columns += list(traffic_df.select_dtypes(['float64']).columns)
print(columns)

for column in columns:
    if column == 'target':
        continue
    traffic_df[column] = np.log1p(traffic_df[column])

traffic_df.head()

traffic_df.to_csv('/content/gdrive/MyDrive/CIC-Darknet/darknet_log.csv', index=False)
